reset content in test answer component
